Tone_ai/requirements.txt
streamlit==1.35.0
numpy
opencv-python
scikit-learn
Pillow
scikit-image
